web/happy_farmer_2/requirements.txt
flask==1.1.2
Pillow==8.1.0
